requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-generativeai>=0.4.1
python-dotenv>=1.0.0
firebase-functions>=0.5.0
firebase-admin>=6.6.0
